// pop-web/src/app/shared/openapi/model/treatment-response-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CodedConcept } from './coded-concept';


export interface TreatmentResponseCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who\'s treatment response is asseessed
     */
    caseId: string;
    /**
     * Clinically-relevant date of the treatment response assessment
     */
    date: string;
    /**
     * The classification of the treatment response according to RECIST
     */
    recist: CodedConcept;
    recistInterpreted?: boolean | null;
    /**
     * Method used to assess and classify the treatment response
     */
    methodology: CodedConcept;
    /**
     * References to the neoplastic entities that were assesed for treatment response
     */
    assessedEntitiesIds?: Array<string>;
    assessedBodysites?: Array<CodedConcept> | null;
}